Virtual Key for Your Repositories: LockedMe
Virtual Key for Your Repositories
1.	Project objective: 
Complete the features of the application by planning the development in terms of sprints and then push the source code to the GitHub repository. As this is a prototyped application, the user interaction will be via a command line. 
2.	Background of the problem statement:
Name of the client: Company Lockers Pvt. Ltd.
Name of the Application: LockedMe.com.
Deadline: 15 Working days (3 weeks).
3.	AIM: To develop a prototype of the application (LockedMe.com). The client aim to digitize their products and chose LockedMe.com. This prototype of the application will be then presented to the relevant stakeholders for the budget approval.

4.	Feature of this application: 
1)	Option to create a user specified file to the application 
2)	 Option to delete a user specified file from the application 
3)	 Option to search a user specified file from the application
4)	 Option to show the list of all the file in the folder from the application